Resolving errors pertaining to "floating_ip" clientgen (#483)
* Resolving errors pertaining to "floating_ip" clientgen

Was running openapi generator and ran into the following exception:
Exception: 'null' defines discriminator 'type', but the referenced schema 'floating_ip_action_unassign' is incorrect. invalid type for type, set it to string

Solution: must specify the discriminator mapping in the base schema, which in this case is "floating_ip_action_type"

Was able to move past that exception and on to another, which will resolved in a different pr.

* cleaning up "type" definitions

removing duplications of type definitions and maintianing it's example and description.