Resolving errors with openapi-generator 5.0.0
When inputting the DO openapi spec into the latest release of openapi-generator, I would get an exception with new_volume_ext4 and new_volume_xfs (a "filesystem_type" needs to be required and a type of string")

Solution: include discriminator in the base schema (volume_write_file_system_type).

This resolved the exceptions.

With this PR, the DO spec is accepted into the latest version of openapi-generator (5.0.0) exception/error free and is able to generate a python client.